+++ catsim/Phantom_Analytic_Get.py
@@ -1,7 +1,7 @@
 # Generated with SMOP  0.41
 # Phantom_Analytic_Get.m
 
-    # -----------------------------------------------------------------------
+ # -----------------------------------------------------------------------
 #   Program Name: Phantom_Analytic_Get.m                                   
 #   Authors:  Samit Basu, Bruno De Man, Jed Pack (GE Global Research)
 #   Organization: GE Global Research and GE Healthcare, General Electric Company
@@ -9,10 +9,10 @@
 #   Date:  Feb 3, 2015
 #   Class: GE Confidential. General Electric Proprietary Data (c) 2012 General Electric Company
     
-    # Aim
+# Aim
 #   Read and parse a phantom from file, store it in C memory, and extract the materials
     
-    # Inputs
+# Inputs
 #   cfg.phantom_filename     The phantom file in one of the following formats:
 #      .pp format            The original FORBILD format used by CatSim 1.0 and 2.0
 #      .ppm format           A FreeMat script that defines the phantom as a cell array of structures
@@ -22,16 +22,16 @@
 #                            The file must also contain a material list such as this:
 #                              materialList = {'water' 'lung' 'bone'};
     
-    #   cfg.phantom_scale        A scale factor to resize phantom,
+#   cfg.phantom_scale        A scale factor to resize phantom,
 #                            also used to convert from cm to mm (cfg.phantom_scale=10).
     
-    #   cfg.time                 (optional) only used for certain dynamic phantoms
+#   cfg.time                 (optional) only used for certain dynamic phantoms
     
-    # Outputs
+# Outputs
 #   materialList             Cell array of phantom materials (strings)
 #                            Note that this is set when the phantom file is "sourced".
     
-    # History: 
+# History: 
 #   2012-10-09 Paul FitzGerald (GE Global Research)
 #              Renamed - was ParsePhantomZZ.
 #              Cleaned up and added "Verbose" output.
@@ -54,9 +54,6 @@ def Phantom_Analytic_Get(cfg):
     Scale=cfg.phantom.scale
     print('Reading phantom file {}...'.format(PhantomFilename))
 
-    #NOTE by jiayong: following two should be a list of dictionaries
-    #phantObject = []
-    
     PhantomPath = os.path.dirname(PhantomFilename)
     BaseName = os.path.basename(PhantomFilename)
     PhantomBaseName, PhantomExtension = BaseName.split('.')
@@ -74,7 +71,6 @@ def Phantom_Analytic_Get(cfg):
     
     print('Reading phantom file {}.'.format(ppmPhantomFilename))
    
-    #TODO: how to source file in xcist?, ask mingye how to do this part
     # phobject is just a temporary var, its values will be saved to objs and passed to setobject
     phobject = parse_analytical_ppm(ppmPhantomFilename)
   
@@ -98,25 +94,29 @@ def Phantom_Analytic_Get(cfg):
     #                if all(abs(object.clip[ii](jj,arange(1,3))) == concat([1,0,0])):
     #                    object.clip[ii][jj,arange(1,3)]=dot(object.clip[ii](jj,arange(1,3)),transmat)
     #
-    ## phantom position offset, Mingye
-    #if isfield(cfg,'phantom_position_offset') and any(cfg.phantom_position_offset) != 0:
-    #    cfg.phantom_position_offset_analytic = copy(cfg.phantom_position_offset)
-    #
-    #if isset('cfg.phantom_position_offset_analytic') and any(cfg.phantom_position_offset_analytic) != 0:
-    #    cfg.phantom_position_offset_analytic = copy(ravel(cfg.phantom_position_offset_analytic).T)
-    #    for ii in arange(1,length(object.type)).reshape(-1):
-    #        object.center[ii,arange()]=object.center(ii,arange()) + cfg.phantom_position_offset_analytic
-    #        if logical_not(isempty(object.clip[ii])):
-    #            # Mingye Wu, Sept 15 2017
-    #            # get the delta D of clip distance
-    #            for jj in arange(1,size(object.clip[ii],1)).reshape(-1):
-    #                d=get_clip_dD(object.clip[ii](jj,arange(1,3)),cfg.phantom_position_offset_analytic)
-    #                object.clip[ii][jj,4]=object.clip[ii](jj,4) + d
-    if len(phobject['materialList']) < max(phobject['material'])[0]:
+    # phantom position offset, Mingye
+    if hasattr(cfg.phantom,'centerOffset') and not np.allclose(cfg.phantom.centerOffset, 0):
+        setattr(cfg.phantom, "centerOffset_analytic", cfg.phantom.centerOffset)
+    
+    if hasattr(cfg.phantom, 'centerOffset_analytic') and not np.allclose(cfg.phantom.centerOffset_analytic, 0):
+        #cfg.phantom_position_offset_analytic = copy(ravel(cfg.phantom_position_offset_analytic).T)
+        for ii in range(len(phobject['type'])):
+            phobject['center'][ii] = np.array(phobject['center'][ii])
+            phobject['center'][ii] += cfg.phantom.centerOffset_analytic
+            if len(phobject['clip'][ii]) > 0:
+                # Mingye Wu, Sept 15 2017
+                # get the delta D of clip distance
+                for jj in range(len(phobject['clip'][ii])):
+                    #breakpoint()
+                    d = get_clip_dD(phobject['clip'][ii][jj][:3], cfg.phantom.centerOffset_analytic)
+                    phobject['clip'][ii][jj][3] += d
+    #breakpoint()
+
+    if len(phobject['materialList']) < max(phobject['material']):
         sys.exit('Phantom cannot reference a material that is not in the list')
     objs = {'params': [[] for _ in range(len(phobject['type']))],
             'clip': [[] for _ in range(len(phobject['type']))]}
-    if np.any(np.array(phobject['type'])[:,0]==100):
+    if np.any(np.array(phobject['type'])==100):
         #TODO: not do this for now
         #Phantom_Polygonal(object,Scale)
         pass
@@ -138,23 +138,25 @@ def Phantom_Analytic_Get(cfg):
         indsToScale= list(range(6)) + [14,15]
 
         for i in range(len(phobject['type'])):
-            if np.logical_not(np.any(phobject['type'][i][0] in [3,7])):
-                tmp = phobject['center'][i] + phobject['half_axes'][i] + phobject['euler_angs'][i] + phobject['density'][i] + phobject['type'][i] + [0] + phobject['shape'][i] + phobject['axial_lims'][i] + phobject['material'][i]
+            if not phobject['type'][i] in [3,7]:
+                tmp = list(phobject['center'][i]) + phobject['half_axes'][i] + phobject['euler_angs'][i] + [phobject['density'][i]] + [phobject['type'][i]] + [0] + [phobject['shape'][i]] + phobject['axial_lims'][i] + [phobject['material'][i]]
                 tmp = np.array(tmp)
                 tmp[indsToScale] = tmp[indsToScale]*Scale
                 #breakpoint()
-                objs['params'][i]=tmp
-                tmp=phobject['clip'][i]
+                objs['params'][i] = tmp
+                #breakpoint()
+                tmp2 = np.array(phobject['clip'][i])
                 #if np.logical_not(np.isempty(tmp)):
-                if len(tmp)!=0:
-                    tmp[:,3]= tmp[:,3]*Scale
-                objs['clip'][i]=tmp
+                if len(tmp2)!=0:
+                    tmp2[:,3]= tmp2[:,3]*Scale
+                objs['clip'][i] = tmp2
             else:
-                objs['params'][i]= np.vstack([phobject['center'][i],phobject['half_axes'][i],phobject['euler_angs'][i],phobject['density'][i],phobject['type'][i],0,phobject['shape'][i],phobject['axial_lims'][i],phobject['material'][i]])
-                objs['clip'][i]=phobject['clip'][i]
+                objs['params'][i] = np.vstack([phobject['center'][i],phobject['half_axes'][i],phobject['euler_angs'][i],phobject['density'][i],phobject['type'][i],0,phobject['shape'][i],phobject['axial_lims'][i],phobject['material'][i]])
+                objs['clip'][i] = phobject['clip'][i]
         # Set the objects, and pass to C.
         objs['params'] = np.array(objs['params'])
         #objs['clip'] = np.array(objs['clip'])
+        #breakpoint()
         phantObject, numObjects, T, cumCP, NumCP, materialIndex, X, K, Q, Eta, S, D = Phantom_Analytic_SetObjects(objs)
         objs['params'] = Phantom_Analytic_ConvertTori(phantObject,objs['params'],objs['clip'])
         phantObject = Phantom_Analytic_BoundObjects(phantObject,objs['params'])
@@ -173,25 +175,26 @@ def Rmat(p=None):
             [np.sin(p2)*np.sin(p3), -np.sin(p2)*np.cos(p3), np.cos(p2)])
     return R
     
-# not used for now    
-#def get_clip_dD(c0=None,c1=None,*args,**kwargs):
-#    varargin = get_clip_dD.varargin
-#    nargin = get_clip_dD.nargin
-#
-#    # Mingye Wu, Sept 15 2017
-## get the delta D of clip
-## c0: original clip vector
-## c1: phantom off-centering coordinates
-#    
-#    r0=norm(c0,2)
-#    r1=norm(c1,2)
-#    d=norm(c0 - c1,2)
-#    cosA=(r0 ** 2 + r1 ** 2 - d ** 2) / (dot(dot(2,r0),r1))
-#    
-#    cosA[logical_or(isnan(cosA),cosA) == logical_or(inf,cosA) == - inf]=0
-#    delta=dot(r1,cosA)
-#    
-#    return delta
+# Mingye Wu, Sept 15 2017
+# get the delta D of clip
+# c0: original clip vector
+# c1: phantom off-centering coordinates
+def get_clip_dD(c0=None,c1=None):
+    
+    c0 = np.array(c0)
+    c1 = np.array(c1)
+    r0 = np.linalg.norm(c0)
+    r1 = np.linalg.norm(c1)
+    d = np.linalg.norm(c0 - c1)
+    cosA = (r0**2 + r1**2 - d**2)/(2*r0*r1)
+    
+    cosA = np.nan_to_num(cosA)
+    #breakpoint()
+    #cosA[cosA==np.inf] = 0.
+
+    delta = r1*cosA
+    
+    return delta
     
 if __name__ == '__main__':
     pass
